# openapi.yaml
openapi: 3.0.3
info:
  title: 千界花园 (Thousand Realms Garden) API
  description: '群智协同操作系统 REST API — XYZ 三轴自由连接、多 Agent 编排、统一 LLM 适配


    📊 自动扩展: 145 个新端点已从后端路由同步'
  version: 1.0.0
  contact:
    name: Thousand Realms Garden Team
  license:
    name: MIT
servers:
- url: http://localhost:3001/api
  description: 本地开发环境
- url: https://thousand-realms-garden.onrender.com/api
  description: 生产环境 (Render)
...
paths:
...
  /groups:
    get:
      tags:
      - Groups
      summary: 获取群组列表
      security:
      - BearerAuth: []
      responses:
        200:
          description: 群组列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
    post:
      tags:
      - Groups
      summary: 创建群组编排
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - mode
              - agents
              properties:
                name:
                  type: string
                description:
                  type: string
                mode:
                  type: string
                  enum:
                  - sequential
                  - parallel
                  - hierarchical
                agents:
                  type: array
                  items:
                    type: string
                  description: Agent ID 列表
      responses:
        201:
          description: 创建成功
  /groups/{id}/execute:
    post:
      tags:
      - Groups
      summary: 执行群组编排
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - input
              properties:
                input:
                  type: string
                  description: 群组输入消息
      responses:
        200:
          description: 执行结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      mode:
                        type: string
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            agentId:
                              type: string
                            agentName:
                              type: string
                            response:
                              type: string
                            latency:
                              type: number
...
  /groups/{id}:
    get:
      tags:
      - Groups
      summary: GET /groups/{id}
      operationId: get_groups_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
      - Groups
      summary: PUT /groups/{id}
      operationId: put_groups_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
      - Groups
      summary: DELETE /groups/{id}
      operationId: delete_groups_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /groups/{id}/agents:
    post:
      tags:
      - Groups
      summary: POST /groups/{id}/agents
      operationId: post_groups_id_agents
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /groups/{id}/agents/{agentId}:
    delete:
      tags:
      - Groups
      summary: DELETE /groups/{id}/agents/{agentId}
      operationId: delete_groups_id_agents_agentId
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: agentId
        in: path
        required: true
        schema:
          type: string
  /groups/{id}/coordinator:
    post:
      tags:
      - Groups
      summary: POST /groups/{id}/coordinator
      operationId: post_groups_id_coordinator
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /groups/{id}/nest:
    post:
      tags:
      - Groups
      summary: POST /groups/{id}/nest
      operationId: post_groups_id_nest
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/chariot/{id}/stats:
    get:
      tags:
      - Handoff
      summary: GET /handoff/chariot/{id}/stats
      operationId: get_handoff_chariot_id_stats
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/initiate:
    post:
      tags:
      - Handoff
      summary: POST /handoff/initiate
      operationId: post_handoff_initiate
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /handoff/list:
    get:
      tags:
      - Handoff
      summary: GET /handoff/list
      operationId: get_handoff_list
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /handoff/{id}/accept:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/accept
      operationId: post_handoff_id_accept
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/cancel:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/cancel
      operationId: post_handoff_id_cancel
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/complete:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/complete
      operationId: post_handoff_id_complete
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/execute:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/execute
      operationId: post_handoff_id_execute
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/reject:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/reject
      operationId: post_handoff_id_reject
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/start:
    post:
      tags:
      - Handoff
      summary: POST /handoff/{id}/start
      operationId: post_handoff_id_start
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /handoff/{id}/status:
    get:
      tags:
      - Handoff
      summary: GET /handoff/{id}/status
      operationId: get_handoff_id_status
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /monitor/groups:
    get:
      tags:
      - Monitoring
      summary: GET /monitor/groups
      operationId: get_monitor_groups
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
...
components:
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        mode:
          type: string
          enum:
          - sequential
          - parallel
          - hierarchical
        agents:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - name
      - mode
      - agents
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 通过 /auth/login 获取的 JWT Token